+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-14T20:43:58.456Z"
+    created: Optional[str] = "2026-07-15T06:24:27.672Z"
 
-    updated: Optional[str] = "2026-07-14T20:43:58.456Z"
+    updated: Optional[str] = "2026-07-15T06:24:27.672Z"
 
     updated_by_id: OptionalNullable[str] = UNSET