+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-14T20:43:58.456Z"
+    created: Optional[str] = "2026-07-15T06:24:27.672Z"
 
-    updated: Optional[str] = "2026-07-14T20:43:58.456Z"
+    updated: Optional[str] = "2026-07-15T06:24:27.672Z"
 
     updated_by_id: OptionalNullable[str] = UNSET