+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsetypescript.py
@@ -40,9 +40,9 @@ class EvaluatorResponseTypescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-14T20:43:58.456Z"
+    created: Optional[str] = "2026-07-15T06:24:27.672Z"
 
-    updated: Optional[str] = "2026-07-14T20:43:58.456Z"
+    updated: Optional[str] = "2026-07-15T06:24:27.672Z"
 
     updated_by_id: OptionalNullable[str] = UNSET