+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXH5Z3PA4PEMZERBXW6857C7"
+        "01KXJ764GCMTTBR7RNEJ1QA3PD"
     )
     r"""The unique identifier of the data source"""