+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXH5Z3PCV3HWJSWA0253WM55"
+        "01KXJ764GE85AXCSJ0Q8GK45Z1"
     )
     r"""The unique identifier of the data source"""