+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXH5Z3PGSXCP84V64JYQ00V4"
+        "01KXJ764GJ22A8B9QWWWXVNKGT"
     )
     r"""The unique identifier of the data source"""